Create rally report dir in root before running Rally, Tempest
If the lab was deployed in the pipeline, the /root does not
contain the /root/rally_reports/ directroy. But the /root is
mounted before running tests and everywhere there is a hard-code
that repors are saved in rally_reports/ folder.

If the /root/rally_reports/ dir already exists (and even contains
some files) - it is not overwritten and the files inside are not
deleted.

Change-Id: I22ad423580d2f7cd08ac8b619cb39e0e60400d95