Fixed the function: remove test docker container by name
Earlier we removed the container by grepping the image link.
The command was not nice and with pepper the awk and single
quotation it is more compicated. Better to remove Rally container
by its name. The 'runRallyScenarios' functions sets the name
for the container. No need to delete container for Tempest, because
docker runs with --rm.

Change-Id: Iccf62ac054315513e28983e0558556ada05af5b3