waitForInputThenPass helper
We need to have a timeout in user input and continue pipeline.
However default behaviour is to fail the job. Unfortunately there
is no easy way to distinguish user abort from timeout. Official
workaround is to look at the user name. But getUser() function
is forbidden by Script Security Plugin. This is why we have to
move this function to pipeline library.

Change-Id: Ic8cfe939f3d2694db49b3413eb246c89f89f9f4f