Mount artifact-metadata dir to /workspace
Sometimes artifact-metadata dir is located under long
enouth path, and being mounted 'as is' it leads to tox
failure. This is because pip (?) replaces shebang string in
'metadata-app' when installs it into virtualenv, and if the
string is long enough we've got failure.

There is github issue related:
  https://github.com/pypa/virtualenv/issues/596

This commit mounts any external path to internal '/workspace'
dir and fixes the issue.

Related-Prod: PRODX-2782
Change-Id: Ie3405086a94a9f7315ce0ba935d1dd68a85d0325