Retry dist-upgrades
If the 'apt-get dist-upgrade' operations running long on the minions in batch
salt master may return an empty string for some minions. The pipelines do not
expectd this and threat it as an error. This patch adds a retry which one one
hand will succeed on the second run since the operation will be much shorter,
on the other hand it will not hide the real issue if it happens.

Closes-Bug: PROD-35355
Change-Id: I95d4c33bf7e28b369f5dc3f468944be15a5d5d6a