Fix node rebooting procedure
Pepper is not capable of returning an error in case if a minion was
not reachable during the task run. It returns "false" in this case
which conflicts with the boolean checks. This commit changes the way
how file existence is checked, removing false positives that come from
unreachable minions.

Closes-Bug: PROD-35442
Change-Id: I6920ef3b08b948582df101ce7c080c1bd282c458